CVE-s/todo-list-app/requirements.txt
Flask
Jinja2
redis
MarkupSafe
python-dotenv